Moving controller logic into modules for easier overriding downstream
